+++ dspic33ch_totem_pole_single_phase_primary.X/revision.h
@@ -33,7 +33,7 @@
 
 #define REV_MAJOR       '1'
 #define REV_MINOR       '1'
-#define REV_PATCH       '2'
+#define REV_PATCH       '3'
 
 /*******************************************************************************
  * @fn      revision history 1.1.2